Refactor tooling and update template
This defines a GitHub Action to test the plugin code against the
security aspects of the WordPress Coding Standards, for the versions of
PHP which we currently use (7.2 and 7.4).

Along the way it also fixes some parts of the plugin tooling, like some
parts of composer.json. It also adds a "lint" script to Composer so that
developers can apply the full suite of coding standards while working
locally.